app/requirements.txt
# AI 모델 관련
# torch와 torchvision은 별도 설치
numpy==1.24.3
opencv-python-headless>=4.7.0
pillow>=9.5.0
pandas>=2.0.0
GPUtil>=0.4.0
ultralytics>=8.0.0

# AI 서버 프레임워크
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# 데이터베이스
sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.6  # SQLite 사용으로 주석 처리
alembic>=1.11.0
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]>=3.3.0

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# 로깅
loguru>=0.7.0 


